pyproject.toml
[tool.poetry]
name = "netbox-nornir"
version = "0.0.1"
description = ""
authors = ["OpticoreIT <opensource@opticoreit.com>"]
readme = "README.md"
packages = [{include = "netbox_nornir"}]

[tool.poetry.dependencies]
python = "^3.8"
nornir = "^3.3.0"
nornir-napalm = "^0.3.0"
netutils = "^1.4.1"
netmiko = "^4.1.2"
nornir-netmiko = "^0"
netbox-celery = "^0.1.1b3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.1"
invoke = "^2.0.0"
pylint-django = "^2.5.3"
bandit = "^1.7.5"
mkdocs = "^1.4.2"
mkdocs-version-annotations = "^1.0.0"
mkdocs-include-markdown-plugin = "3.8"
pymdown-extensions = "^9.10"
mkdocs-material = "^9.1.4"
mkdocstrings = "^0.20.0"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
        /(
          \.git
        | \.tox
        | \.venv
        | env/
        | _build
        | build
        | dist
        | netbox/*
        )/ 
    '''

[tool.pylint.master]
load-plugins = "pylint_django,pylint_celery"
django-settings-module = "netbox.settings"
ignore-paths = [
    ".*netbox/.*",
    ".*third_party.*",
    ".*migrations.*",
    ".*tests.*",
    ".*example.py",
    ".*configuration.py",
    "tasks.py",
]

[tool.pylint.messages_control]
# TODO: re-enable and fix these as time permits
# unused-import is already covered by flake8
disable=""",
    duplicate-code,
    line-too-long,
    super-init-not-called,
    super-with-arguments,
    too-few-public-methods,
    too-many-ancestors,
    too-many-arguments,
    too-many-boolean-expressions,
    too-many-branches,
    too-many-instance-attributes,
    too-many-lines,
    too-many-locals,
    too-many-nested-blocks,
    too-many-public-methods,
    too-many-return-statements,
    too-many-statements,
    wrong-import-order,
    wrong-import-position,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
